refactor and add cli
